requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.4.0
djangorestframework-simplejwt[blacklist]==5.3.1
pytest
pytest-django
python-dotenv==1.0.1
pytest-cov==5.0.0
black==24.8.0
isort==5.13.2